requirements.txt
discord.py==2.4.0
Flask==3.1.0
joblib==1.3.2
matplotlib==3.8.2
pandas==2.2.3
prophet==1.1.6
python-dotenv==1.0.1
Requests==2.32.3
yfinance==0.2.55
